# Conflicts:
#	pom.xml
#	src/main/java/com/gameloft9/demo/service/impl/system/FinancePurchaseBillPayServiceImpl.java
#	src/main/java/com/gameloft9/demo/service/impl/system/LoginServiceImpl.java
#	src/main/java/com/gameloft9/demo/utils/NumberUtil.java
#	src/main/resources/applicationContext.xml
#	src/main/resources/config/config_uat.properties
#	src/main/webapp/layuicms/common/js/api.js
#	src/main/webapp/layuicms/index.js
#	src/main/webapp/layuicms/page/system/purOrder/purApplyAdd.html
#	src/main/webapp/layuicms/page/system/purOrder/purApplyAdd.js
#	src/main/webapp/layuicms/page/system/purOrder/purApplyList.html

+++ src/main/java/com/gameloft9/demo/service/impl/system/LenProductServiceImpl.java
@@ -4,9 +4,14 @@ import com.gameloft9.demo.dataaccess.dao.system.LenProductMapper;
 import com.gameloft9.demo.dataaccess.model.system.LenProduct;
 import com.gameloft9.demo.service.api.system.LenProductService;
 import com.gameloft9.demo.service.beans.system.PageRange;
+import com.gameloft9.demo.utils.Constants;
+import com.gameloft9.demo.utils.DocumentNumberUtil;
 import com.gameloft9.demo.utils.UUIDUtil;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -17,6 +22,7 @@ import java.util.List;
  * @description:
  */
 @Service
+@Transactional(rollbackFor = Exception.class)
 public class LenProductServiceImpl implements LenProductService {
 
     @Autowired
@@ -42,37 +48,51 @@ public class LenProductServiceImpl implements LenProductService {
 
     @Override
     public boolean insert(LenProduct lenProduct) {
+        String number = DocumentNumberUtil.getDocumentNumber("P");
         String uuid = UUIDUtil.getUUID();
         LenProduct product = new LenProduct();
         product.setId(uuid);
         product.setProductName(lenProduct.getProductName());
+        product.setProductNumber(lenProduct.getProductNumber());
+        product.setCanSold(lenProduct.getCanSold());
+        product.setSupportPrice(lenProduct.getSupportPrice());
         product.setProductType(lenProduct.getProductType());
         product.setProductDescribe(lenProduct.getProductDescribe());
-        product.setState(lenProduct.getState());
-        product.setWasteId(lenProduct.getWasteId());
-        if (mapper.insert(product)>0){
+        product.setProductState(Constants.productState.UN_TIJIAO);
+        product.setOther1(number);
+        if (mapper.insert(product) > 0) {
             return true;
-        }else {
-            return  false;
+        } else {
+            return false;
         }
     }
 
     @Override
     public boolean update(LenProduct lenProduct) {
-        if (mapper.update(lenProduct)>0){
+        String sysUser = (String) SecurityUtils.getSubject().getSession().getAttribute("sysUser");
+        if(lenProduct.getCanSold().equals(sysUser)){
+        lenProduct.setProductState(Constants.productState.UN_TIJIAO);
+        if (mapper.update(lenProduct) > 0) {
             return true;
-        }else {
-            return  false;
+        } else {
+            throw new RuntimeException(">>>>>>>数据库操作失败<<<<<<<");
+        }
+
+        }else{
+            throw new RuntimeException(">>>>>>>用户信息不匹配，不能操作<<<<<<<");
         }
     }
 
     @Override
     public boolean delete(String id) {
-        if (mapper.delete(id)>0){
-            return true;
+        if(Constants.lennonPDAudi()==1){
+           if (mapper.delete(id)>0){
+               return true;
+           }
         }else {
-            return  false;
+            return false;
         }
+        return false;
     }
 
     @Override
@@ -80,14 +100,159 @@ public class LenProductServiceImpl implements LenProductService {
         return mapper.dataCount(state);
     }
 
+
     @Override
-    public boolean changeState() {
-        if (mapper.changeState()>0){
-            return true;
-        }else {
-            return  false;
+    public boolean stepBack(String id) {
+        /**
+         * 只有在未审核（包括）之前的状态才能进行撤回操作
+         */
+
+        LenProduct len = mapper.getByPrimaryKey(id);
+        String state = len.getProductState();
+        if (Constants.TIJIAO.equals(state)){
+            if (mapper.changeState(Constants.UN_TIJIAO,id)>0) {
+                return true;
+            }else{
+                return false;
+
+            }
+        }else if (Constants.UN_AUDI.equals(state)){
+            if (mapper.changeState(Constants.TIJIAO,id)>0){
+                return true;
+            }else {
+                return false;
+            }
+        }else{
+            return false;
         }
+
     }
+
+    /**
+     * 更改产品状态
+     *
+     * @param
+     * @return
+     */
+    @Override
+    public boolean changeState(String id) {
+        LenProduct product = mapper.getByPrimaryKey(id);
+        String state = product.getProductState();
+        Subject subject = SecurityUtils.getSubject();
+        //未提交————提交 1-2
+        if (state.equals(Constants.productState.UN_TIJIAO)){
+            mapper.changeState(Constants.productState.TIJIAO_UNAUDI,id);
+            return true;
+        }
+        //领料未分配--分配生产 4-5
+        if (state.equals(Constants.productState.REACH_UNFENPEI)){
+            if (subject.hasRole(Constants.PRODUCE_ADMIN)){
+            mapper.changeState(Constants.productState.FENPEI_START_PRODUCE,id);
+            return true;}else{
+                return false;
+            }
+        }
+                //审核未领料---领料待分配 3-4（华峰授权步骤）
+            if (product.getProductState().equals(Constants.productState.AUDI_UNREACH)) {
+                if (subject.hasRole(Constants.PRODUCE_ADMIN)) {
+                    mapper.changeState(Constants.productState.REACH_UNFENPEI, id);
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+            //提交审核 2-3
+            if (product.getProductState().equals(Constants.productState.TIJIAO_UNAUDI)){
+                if (subject.hasRole(Constants.PRODUCE_ADMIN)) {
+                     mapper.changeState(Constants.productState.AUDI_UNREACH, id);
+                     return true;
+                } else {
+                    return false;
+                     }
+
+                }
+
+        return false;
+        }
+
+
+        @Override
+        public boolean changeBehindState(String state, String id){
+           if (mapper.changeState(state,id)>0){
+               return true;
+           }else {
+               return false;
+           }
+
+
+        }
+         @Override
+         public boolean insertSupportPrice(String supportPrice, String id) {
+            if (mapper.insertSupportPrice(supportPrice,id)>0){
+                return true;
+            }else {
+                return false;
+            }
+    }
+
+    /*当前状态为 0*/
+        /*if (Constants.UN_TIJIAO.equals(state)) {
+
+            if (mapper.changeState(Constants.TIJIAO, id) > 0) {
+                return true;
+            } else {
+
+                return false;
+            }
+            *//*当前状态 state：1*//*
+        } else if (Constants.TIJIAO.equals(state) ) {
+
+            if (mapper.changeState(Constants.UN_AUDI, id) > 0) {
+                return true;
+            } else {
+                return false;
+            }
+
+            *//*state:2*//*
+        } else if (Constants.UN_AUDI.equals(state)) {
+            if (subject.hasRole(Constants.PRODUCE_ADMIN)) {
+                if (mapper.changeState(Constants.ADUI, id) > 0) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+
+        } else if (Constants.ADUI.equals(state)) {
+            if (subject.hasRole(Constants.PRODUCE_ADMIN)) {
+                if (mapper.changeState(Constants.PRODUCING, id) > 0) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        } else if (Constants.PRODUCING.equals( state)) {
+            if (subject.hasRole(Constants.PRODUCE_ADMIN)) {
+                if (mapper.changeState(Constants.INTO_DEPOT, id) > 0) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+
+
+    }else {
+        return false;
+    }
+
+
+
+            return false;*/
+
+
+
+
+
 
     @Override
     public boolean updateByPrimaryKeySelective(LenProduct lenProduct) {
@@ -105,5 +270,73 @@ public class LenProductServiceImpl implements LenProductService {
         }else {
             return  false;
         }
+    }
+
+    @Override
+    public boolean changeProState(String state, String id) {
+       if (mapper.changeState(state,id)>0){
+           return true;
+       }else {
+          return false;
+       }
+    }
+
+    @Override
+    public List<LenProduct> selectByState() {
+        return mapper.selectByState();
+    }
+
+    /**
+     * 获取产品id
+     * */
+    @Override
+    public List<String> getProductId() {
+        return mapper.getProductId();
+    }
+
+
+
+
+
+    @Override
+    public boolean
+    huaOutDepot(String id) {
+        //华峰出库
+       if (mapper.changeState(Constants.productState.DEPOT_OUT,id)>0){
+           return true;
+       }else {
+
+
+
+           return false;
+       }
+
+    }
+
+    @Override
+    public boolean huaInDepot(String id) {
+        if (mapper.changeState(Constants.productState.INTO_DEPOT,id)>0){
+            return true;
+        }else{
+
+
+            return false;
+        }
+    }
+
+    @Override
+    public boolean productIntoDepot(String id) {
+        if (mapper.changeState(Constants.productState.INTO_DEPOT,id)>0){
+            return true;
+        }else {
+            return false;
+        }
+    }
+
+    //插入到仓库
+    @Override
+    public boolean inssrDepot(LenProduct lenProduct) {
+
+        return false;
     }
 }
